+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - ``RobotWrenchDrawer`` became ``RobotDiscWrenchDrawer`` (discrete velocity differentiation)
+- InvertedPendulum: clamping is now the default behavior for ``set_cop`` and ``set_lambda``
 
 ## [1.1.0] - 2019/04/30